# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "This is the API documentation for my Invoices application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/tax-profile/list": {
            "post": {
                "tags": [
                    "TaxProfile"
                ],
                "summary": "List Tax Profiles",
                "description": "Returns a list of tax profiles based on provided filters, sort, and limit parameters. Requires authentication.",
                "operationId": "4c02c5b06942f82e53369593b3edffc6",
                "requestBody": {
                    "description": "Filtering, sorting, and pagination parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaginatedListFilter"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of tax profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TaxProfile"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tax-profile/{id}": {
            "get": {
                "tags": [
                    "TaxProfile"
                ],
                "summary": "Get Tax Profile",
                "description": "Retrieve a specific tax profile by its ID. Requires authentication.",
                "operationId": "f2cc5760291316bd02ec96d3532bdc7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tax profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tax profile found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaxProfile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tax Profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tax Profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "TaxProfile"
                ],
                "summary": "Update Tax Profile",
                "description": "Updates an existing tax profile. Requires authentication.",
                "operationId": "383353ac1539a7d3f01588629da0f8d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tax profile to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tax profile data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tax profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaxProfile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tax Profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tax Profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TaxProfile"
                ],
                "summary": "Delete Tax Profile",
                "description": "Deletes a tax profile by its ID. Requires authentication.",
                "operationId": "be5bcf1c389fd5802e0983dfd71751ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tax profile to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tax profile deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tax Profile deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tax Profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tax Profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tax-profile": {
            "post": {
                "tags": [
                    "TaxProfile"
                ],
                "summary": "Create Tax Profile",
                "description": "Creates a new tax profile for the authenticated user.",
                "operationId": "38b2f7da05a2bdba034f828877eed960",
                "requestBody": {
                    "description": "Tax profile details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tax_id",
                                    "company_name",
                                    "address",
                                    "country",
                                    "city",
                                    "zip_code"
                                ],
                                "properties": {
                                    "tax_id": {
                                        "type": "string",
                                        "example": "TAX123456"
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "example": "Acme Inc."
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main Street"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "zip_code": {
                                        "type": "string",
                                        "example": "10001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tax profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaxProfile"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "{ 'tax_id': ['The tax_id field is required.'] }"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Filter": {
                "title": "Filter",
                "description": "A single filter condition.",
                "required": [
                    "field",
                    "operator"
                ],
                "properties": {
                    "field": {
                        "description": "Field name to filter on.",
                        "type": "string",
                        "example": "user_id"
                    },
                    "value": {
                        "description": "Value to filter by.",
                        "type": "string",
                        "example": "36"
                    },
                    "fieldType": {
                        "description": "Type of the field.",
                        "type": "string",
                        "enum": [
                            "text",
                            "date",
                            "number",
                            "boolean",
                            "set",
                            "array"
                        ],
                        "example": "number"
                    },
                    "operator": {
                        "description": "Operator for the filter condition.",
                        "type": "string",
                        "enum": [
                            "contains",
                            "notContains",
                            "equals",
                            "notEqual",
                            "startsWith",
                            "endsWith",
                            "blank",
                            "notBlank",
                            "greaterThan",
                            "greaterThanOrEqual",
                            "lessThan",
                            "lessThanOrEqual",
                            "inRange"
                        ],
                        "example": "equals"
                    },
                    "rangeValue": {
                        "description": "Secondary value for range filters (if applicable).",
                        "type": "string",
                        "example": "50",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginatedListFilter": {
                "title": "Paginated List Filter",
                "description": "Parameters for filtering, sorting, and pagination.",
                "required": [
                    "filters",
                    "sort",
                    "limit"
                ],
                "properties": {
                    "filters": {
                        "description": "Array of filter objects",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Filter"
                        }
                    },
                    "sort": {
                        "$ref": "#/components/schemas/Sort"
                    },
                    "limit": {
                        "description": "Maximum number of records to return",
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
            "Sort": {
                "title": "Sort",
                "description": "Sorting options for a query.",
                "required": [
                    "field",
                    "direction"
                ],
                "properties": {
                    "field": {
                        "description": "Field to sort by.",
                        "type": "string",
                        "example": "invoice_number"
                    },
                    "direction": {
                        "description": "Sort direction.",
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "example": "asc"
                    }
                },
                "type": "object"
            },
...
            "TaxProfile": {
                "title": "Tax Profile",
                "description": "Tax profile associated with a user, containing tax-related information.",
                "required": [
                    "user_id",
                    "tax_id",
                    "company_name",
                    "address",
                    "country",
                    "city",
                    "zip_code"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the tax profile",
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID of the user that owns the tax profile",
                        "type": "integer",
                        "example": 1
                    },
                    "tax_id": {
                        "description": "Tax identification number",
                        "type": "string",
                        "example": "TAX123456"
                    },
                    "company_name": {
                        "description": "Name of the company",
                        "type": "string",
                        "example": "Acme Inc."
                    },
                    "address": {
                        "description": "Company address",
                        "type": "string",
                        "example": "123 Main Street"
                    },
                    "country": {
                        "description": "Country where the company is located",
                        "type": "string",
                        "example": "USA"
                    },
                    "city": {
                        "description": "City where the company is located",
                        "type": "string",
                        "example": "New York"
                    },
                    "zip_code": {
                        "description": "Zip code for the company's location",
                        "type": "string",
                        "example": "10001"
                    },
                    "created_at": {
                        "description": "Timestamp when the tax profile was created",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2022-01-01T12:00:00Z",
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "Timestamp when the tax profile was last updated",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2022-01-02T12:00:00Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}